Update com.github.macadmins.Nudge.plist
Reverted pfm_macos_min to v11

Reviewed Ari's comment rregarding `nudgeMajorUpgradeEventLaunch`. I did a github search on the Nudge repository, and determined that this is not a key, it's a mistake in documentation. `nudgeMajorUpgradeEventLaunchDelay` (notice "Delay") and `nudgeMinorUpdateEventLaunchDelay` are new keys.

These keys already exist in the Nudge manifext, however they were NOT included in `pfm_segments > UX` like other keys of this type. So I added them there.